refactor: extract business logic from compute_provider router into services
The compute_provider.py router was 3,151 lines with business logic mixed
into route handlers. This refactoring follows the project's architecture
rule that business logic belongs in services, not routers.

Changes:
- Create routers/dependencies.py with ProviderContext dataclass and
  get_provider_for_request FastAPI dependency, eliminating 23 repetitions
  of the provider lookup boilerplate
- Create shared/cloud_credentials.py with AWS/GCP/Azure credential setup
  helpers (pure utility functions with no HTTP dependency)
- Create services/compute_provider_service.py with all business logic:
  launch_task, create_sweep_parent_job, launch_sweep_jobs,
  resume_from_checkpoint, and cluster operation wrappers
- Slim down the router from 3,151 to 1,250 lines of pure HTTP handling

No API contract changes — same routes, same request/response shapes.
